refactor copyVote funct as a service
