+++ src/app/dashboard-widget/profile-card/profile-card.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
 import Swal from 'sweetalert2';
+import { FunctionsService } from './../../services/functions.service';
+
 
 @Component({
   selector: 'cdk-profile-card',
@@ -17,42 +19,17 @@ export class ProfileCardComponent implements OnInit {
   @Input() button_length: string;
   @Input() button_click: string;
 
+  constructor(private functionsService: FunctionsService){
+    this.functionsService;
+  }
+
   width:number = 0;
 
-  constructor() { }
-
-  copyMessage(val: string, text:string){
-    let selBox = document.createElement('textarea');
-    selBox.style.position = 'fixed';
-    selBox.style.left = '0';
-    selBox.style.top = '0';
-    selBox.style.opacity = '0';
-    selBox.value = val;
-    document.body.appendChild(selBox);
-    selBox.focus();
-    selBox.select();
-    document.execCommand('copy');
-    document.body.removeChild(selBox);
-    Swal.fire("Success",text,"success");
-  }
-  copyVote(val: string){
-   let selBox = document.createElement('textarea');
-   selBox.style.position = 'fixed';
-   selBox.style.left = '0';
-   selBox.style.top = '0';
-   selBox.style.opacity = '0';
-   selBox.value = val;
-   document.body.appendChild(selBox);
-   selBox.focus();
-   selBox.select();
-   document.execCommand('copy');
-   document.body.removeChild(selBox);
-   Swal.fire("Success","The vote has been copied to the clipboard","success");
-  }
   redirect(url:string){
-  window.location.href = url;
+    window.location.href = url;
   }
   ngOnInit() {
+
     this.width = window.innerWidth * 0.9;
   }
 }
