+++ src/app/delegates_information/delegates_information.component.ts
@@ -3,6 +3,7 @@ import { ActivatedRoute } from '@angular/router';
 import {HttpdataService} from '../services/http-request.service';
 import Swal from 'sweetalert2';
 import { Title } from '@angular/platform-browser';
+import { FunctionsService } from '../services/functions.service';
 
 @Component({
     selector: 'app-delegates_information',
@@ -23,46 +24,33 @@ export class Delegates_informationComponent implements OnInit {
     server_specs:string;
     public_address:string;
 
-    constructor(private route: ActivatedRoute, private httpdataservice: HttpdataService, private titleService:Title) {
+    constructor(private route: ActivatedRoute, private httpdataservice: HttpdataService, private titleService:Title, private functionsService: FunctionsService) {
         this.delegate_name = this.route.snapshot.queryParamMap.get("data");
-        this.titleService.setTitle(   this.delegate_name +  "Delegate Information - Delegates Explorer - X-CASH");
+        this.titleService.setTitle( this.delegate_name +  "Delegate Information - Delegates Explorer - X-CASH" );
+        this.functionsService;
      }
 
 
     ngOnInit() {
 
+
+
       this.httpdataservice.get_request(this.httpdataservice.GET_DELEGATES_INFORMATION + "?parameter1=" + this.delegate_name).subscribe(
-      (res) => {
-                var data = JSON.parse(JSON.stringify(res));
+        (res) => {
+              var data = JSON.parse(JSON.stringify(res));
 
-                this.data = data;
-                this.about = data.about;
-                this.website = data.website;
-                this.team = data.team;
-                this.shared_delegate_status = data.shared_delegate_status == 'true' ? 'Shared'  : 'Solo';
-                this.delegate_fee = data.delegate_fee;
-                this.server_specs = data.server_specs;
-                this.public_address = data.public_address;
-              },
-              (error) =>
-              {
-                Swal.fire("Error","An error has occured.<br/>Get delegates information failed.","error");
-              }
-      );
+              this.data = data;
+              this.about = data.about;
+              this.website = data.website;
+              this.team = data.team;
+              this.shared_delegate_status = data.shared_delegate_status == 'true' ? 'Shared'  : 'Solo';
+              this.delegate_fee = data.delegate_fee;
+              this.server_specs = data.server_specs;
+              this.public_address = data.public_address;
+            },
+            (error) =>    {
+              Swal.fire("Error","An error has occured.<br/>Get delegates information failed.","error");
+            }
+        );
     }
-
-    copyVote(val: string){
-     let selBox = document.createElement('textarea');
-     selBox.style.position = 'fixed';
-     selBox.style.left = '0';
-     selBox.style.top = '0';
-     selBox.style.opacity = '0';
-     selBox.value = val;
-     document.body.appendChild(selBox);
-     selBox.focus();
-     selBox.select();
-     document.execCommand('copy');
-     document.body.removeChild(selBox);
-     Swal.fire("Success","The vote has been copied to the clipboard","success");
-   }
 }
