+++ src/app/services/functions.service.ts
@@ -1,0 +1,26 @@
+import { Injectable } from '@angular/core';
+import Swal from 'sweetalert2';
+
+@Injectable({
+  providedIn: 'root'
+})
+
+export class FunctionsService {
+  constructor() { }
+
+  copyVote(val: string){
+   let selBox = document.createElement('textarea');
+   selBox.style.position = 'fixed';
+   selBox.style.left = '0';
+   selBox.style.top = '0';
+   selBox.style.opacity = '0';
+   selBox.value = val;
+   document.body.appendChild(selBox);
+   selBox.focus();
+   selBox.select();
+   document.execCommand('copy');
+   document.body.removeChild(selBox);
+   Swal.fire("Success","The vote has been copied to the clipboard","success");
+  }
+
+}
