+++ src/app/tables/delegates_voters_list/delegates_voters_list.component.ts
@@ -5,6 +5,7 @@ import { Component, OnInit , ElementRef, ViewChild} from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { ExampleDatabase, ExampleDataSource } from './helpers.data';
 import {HttpdataService} from '../../services/http-request.service';
+import { FunctionsService } from '../../services/functions.service';
 import Swal from 'sweetalert2';
 
 import { MatPaginator, MatSort } from '@angular/material';
@@ -35,7 +36,9 @@ export class Delegates_voters_listComponent implements OnInit {
 	public showFilterTableCode;
   length;
 
-	constructor(private route: ActivatedRoute, private httpdataservice: HttpdataService) { }
+	constructor(private route: ActivatedRoute, private httpdataservice: HttpdataService, private functionsService: FunctionsService) {
+    this.functionsService;
+  }
 
   @ViewChild(MatPaginator) paginator: MatPaginator;
 	@ViewChild(MatSort) sort: MatSort;
@@ -82,21 +85,6 @@ export class Delegates_voters_listComponent implements OnInit {
       }
 	  );
 	}
-
-  copyVote(val: string){
-   let selBox = document.createElement('textarea');
-   selBox.style.position = 'fixed';
-   selBox.style.left = '0';
-   selBox.style.top = '0';
-   selBox.style.opacity = '0';
-   selBox.value = val;
-   document.body.appendChild(selBox);
-   selBox.focus();
-   selBox.select();
-   document.execCommand('copy');
-   document.body.removeChild(selBox);
-   Swal.fire("Success","The vote has been copied to the clipboard","success");
- }
 
 
 }
